+++ mangadlp/app.py
@@ -9,6 +9,47 @@ from mangadlp import downloader, utils
 from mangadlp.api.mangadex import Mangadex
 from mangadlp.cache import CacheDB
 from mangadlp.hooks import run_hook
+from mangadlp.metadata import write_metadata
+from mangadlp.utils import get_file_format
+
+
+def match_api(url_uuid: str) -> type:
+    """
+    Match the correct api class from a string
+
+    Args:
+        url_uuid: url/uuid to check
+
+    Returns:
+        The class of the API to use
+    """
+
+    # apis to check
+    apis: list[tuple[str, re.Pattern, type]] = [
+        (
+            "mangadex.org",
+            re.compile(
+                r"(mangadex.org)|([a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12})"
+            ),
+            Mangadex,
+        ),
+        (
+            "test.org",
+            re.compile(r"(test.test)"),
+            type,
+        ),
+    ]
+
+    # check url for match
+    for api_name, api_re, api_cls in apis:
+        if not api_re.search(url_uuid):
+            continue
+        log.info(f"API matched: {api_name}")
+        return api_cls
+
+    # no supported api found
+    log.error(f"No supported api in link/uuid found: {url_uuid}")
+    raise ValueError
 
 
 class MangaDLP:
@@ -16,18 +57,23 @@ class MangaDLP:
     After initialization, start the script with the function get_manga().
 
     Args:
-        url_uuid (str): URL or UUID of the manga
-        language (str): Manga language with country codes. "en" --> english
-        chapters (str): Chapters to download, "all" for every chapter available
-        list_chapters (bool): List all available chapters and exit
-        file_format (str): Archive format to create. An empty string means don't archive the folder
-        forcevol (bool): Force naming of volumes. Useful for mangas where chapters reset each volume
-        download_path (str/Path): Download path. Defaults to '<script_dir>/downloads'
-        download_wait (float): Time to wait for each picture to download in seconds
-
+        url_uuid: URL or UUID of the manga
+        language: Manga language with country codes. "en" --> english
+        chapters: Chapters to download, "all" for every chapter available
+        list_chapters: List all available chapters and exit
+        file_format: Archive format to create. An empty string means don't archive the folder
+        forcevol: Force naming of volumes. Useful for mangas where chapters reset each volume
+        download_path: Download path. Defaults to '<script_dir>/downloads'
+        download_wait: Time to wait for each picture to download in seconds
+        manga_pre_hook_cmd: Command(s) to before after each manga
+        manga_post_hook_cmd: Command(s) to run after each manga
+        chapter_pre_hook_cmd: Command(s) to run before each chapter
+        chapter_post_hook_cmd: Command(s) to run after each chapter
+        cache_path: Path to the json cache. If emitted, no cache is used
+        add_metadata: Flag to toggle creation & inclusion of metadata
     """
 
-    def __init__(
+    def __init__(  # pylint: disable=too-many-locals
         self,
         url_uuid: str,
         language: str = "en",
@@ -44,6 +90,7 @@ class MangaDLP:
         chapter_pre_hook_cmd: str = "",
         chapter_post_hook_cmd: str = "",
         cache_path: str = "",
+        add_metadata: bool = True,
     ) -> None:
         # init parameters
         self.url_uuid = url_uuid
@@ -60,20 +107,20 @@ class MangaDLP:
         self.manga_post_hook_cmd = manga_post_hook_cmd
         self.chapter_pre_hook_cmd = chapter_pre_hook_cmd
         self.chapter_post_hook_cmd = chapter_post_hook_cmd
-        self.hook_infos: dict = {}
         self.cache_path = cache_path
+        self.add_metadata = add_metadata
+        self.hook_infos: dict = {}
 
         # prepare everything
         self._prepare()
 
     def _prepare(self) -> None:
-        # set manga format suffix
-        if self.file_format and self.file_format[0] != ".":
-            self.file_format = f".{self.file_format}"
+        # check and set correct file suffix/format
+        self.file_format = get_file_format(self.file_format)
         # start prechecks
-        self.pre_checks()
+        self._pre_checks()
         # init api
-        self.api_used = self.check_api(self.url_uuid)
+        self.api_used = match_api(self.url_uuid)
         try:
             log.debug("Initializing api")
             self.api = self.api_used(self.url_uuid, self.language, self.forcevol)
@@ -86,9 +133,9 @@ class MangaDLP:
         # get chapter list
         self.manga_chapter_list = self.api.chapter_list
         self.manga_total_chapters = len(self.manga_chapter_list)
-        self.manga_path = Path(f"{self.download_path}/{self.manga_title}")
+        self.manga_path = self.download_path / self.manga_title
 
-    def pre_checks(self) -> None:
+    def _pre_checks(self) -> None:
         # prechecks userinput/options
         # no url and no readin list given
         if not self.url_uuid:
@@ -112,27 +159,6 @@ class MangaDLP:
             if not self.forcevol and ":" in self.chapters:
                 log.error("Don't specify the volume without --forcevol")
                 raise ValueError
-
-    # check the api which needs to be used
-    def check_api(self, url_uuid: str) -> type:
-        # apis to check
-        api_mangadex = re.compile(
-            r"(mangadex.org)|([a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12})"
-        )
-        api_test = re.compile("test.test")
-
-        # check url for match
-        if api_mangadex.search(url_uuid):
-            log.debug("Matched api: mangadex.org")
-            return Mangadex
-        # this is only for testing multiple apis
-        if api_test.search(url_uuid):
-            log.critical("Not supported yet")
-            raise ValueError
-
-        # no supported api found
-        log.error(f"No supported api in link/uuid found: {url_uuid}")
-        raise ValueError
 
     # once called per manga
     def get_manga(self) -> None:
@@ -166,7 +192,9 @@ class MangaDLP:
 
         # prepare cache if specified
         if self.cache_path:
-            cache = CacheDB(self.cache_path, self.manga_uuid, self.language)
+            cache = CacheDB(
+                self.cache_path, self.manga_uuid, self.language, self.manga_title
+            )
             cached_chapters = cache.db_uuid_chapters
             log.info(f"Cached chapters: {cached_chapters}")
 
@@ -200,23 +228,40 @@ class MangaDLP:
         error_chapters: list[Any] = []
         for chapter in chapters_to_download:
             if self.cache_path and chapter in cached_chapters:
-                log.info("Chapter is in cache. Skipping download")
+                log.info(f"Chapter '{chapter}' is in cache. Skipping download")
                 continue
 
+            # download chapter
             try:
                 chapter_path = self.get_chapter(chapter)
             except KeyboardInterrupt as exc:
                 raise exc
             except FileExistsError:
+                # skipping chapter download as its already available
                 skipped_chapters.append(chapter)
                 # update cache
                 if self.cache_path:
                     cache.add_chapter(chapter)
                 continue
             except Exception:
+                # skip download/packing due to an error
                 error_chapters.append(chapter)
                 continue
 
+            # add metadata
+            if self.add_metadata:
+                try:
+                    metadata = self.api.create_metadata(chapter)
+                    write_metadata(
+                        chapter_path,
+                        {"Format": self.file_format[1:], **metadata},
+                    )
+                except Exception as exc:
+                    log.warning(
+                        f"Can't write metadata for chapter '{chapter}'. Reason={exc}"
+                    )
+
+            # pack downloaded folder
             if self.file_format:
                 try:
                     self.archive_chapter(chapter_path)
@@ -266,7 +311,7 @@ class MangaDLP:
     # once called per chapter
     def get_chapter(self, chapter: str) -> Path:
         # get chapter infos
-        chapter_infos = self.api.get_chapter_infos(chapter)
+        chapter_infos: dict = self.api.manga_chapter_data[chapter]
         log.debug(f"Chapter infos: {chapter_infos}")
 
         # get image urls for chapter