+++ mangadlp/utils.py
@@ -1,7 +1,7 @@
 import re
 from datetime import datetime
 from pathlib import Path
-from typing import Any
+from typing import Any, List
 from zipfile import ZipFile
 
 from loguru import logger as log
@@ -9,7 +9,7 @@ from loguru import logger as log
 
 # create an archive of the chapter images
 def make_archive(chapter_path: Path, file_format: str) -> None:
-    zip_path: Path = Path(f"{chapter_path}.zip")
+    zip_path = Path(f"{chapter_path}.zip")
     try:
         # create zip
         with ZipFile(zip_path, "w") as zipfile:
@@ -29,7 +29,7 @@ def make_pdf(chapter_path: Path) -> None:
         log.error("Cant import img2pdf. Please install it first")
         raise exc
 
-    pdf_path: Path = Path(f"{chapter_path}.pdf")
+    pdf_path = Path(f"{chapter_path}.pdf")
     images: list[str] = []
     for file in chapter_path.iterdir():
         images.append(str(file))
@@ -41,7 +41,7 @@ def make_pdf(chapter_path: Path) -> None:
 
 
 # create a list of chapters
-def get_chapter_list(chapters: str, available_chapters: list) -> list:
+def get_chapter_list(chapters: str, available_chapters: list) -> List[str]:
     # check if there are available chapter
     chapter_list: list[str] = []
     for chapter in chapters.split(","):
@@ -83,7 +83,6 @@ def get_chapter_list(chapters: str, available_chapters: list) -> list:
 
 # remove illegal characters etc
 def fix_name(filename: str) -> str:
-    log.debug(f"Input name='{filename}'")
     filename = filename.encode(encoding="utf8", errors="ignore").decode(encoding="utf8")
     # remove illegal characters
     filename = re.sub(r'[/\\<>:;|?*!@"]', "", filename)
@@ -94,7 +93,7 @@ def fix_name(filename: str) -> str:
     # remove trailing and beginning spaces
     filename = re.sub("([ \t]+$)|(^[ \t]+)", "", filename)
 
-    log.debug(f"Output name='{filename}'")
+    log.debug(f"Input name='{filename}', Output name='{filename}'")
     return filename
 
 
@@ -144,6 +143,20 @@ def get_filename(
         return f"Vol. {chapter_vol} Ch. {chapter_num} - {chapter_name}"
 
     return f"Ch. {chapter_num} - {chapter_name}"
+
+
+def get_file_format(file_format: str) -> str:
+    if not file_format:
+        return ""
+
+    if re.match(r"\.?[a-z0-9]+", file_format, flags=re.I):
+        if file_format[0] != ".":
+            file_format = f".{file_format}"
+    else:
+        log.error(f"Invalid file format: '{file_format}'")
+        raise ValueError
+
+    return file_format
 
 
 def progress_bar(progress: float, total: float) -> None: