+++ tests/test_04_input.py
@@ -1,7 +1,4 @@
 import os
-from pathlib import Path
-
-import pytest
 
 import mangadlp.cli as mdlpinput