+++ tests/test_21_full.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-import mangadlp.app as app
+from mangadlp import app
 
 
 @pytest.fixture
@@ -22,10 +22,12 @@ def wait_20s():
 
 
 def test_full_api_mangadex(wait_20s):
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1.cbz")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.cbz"
+    )
     mdlp = app.MangaDLP(
-        url_uuid="https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie",
+        url_uuid="https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko",
         language="en",
         chapters="1",
         list_chapters=False,
@@ -43,13 +45,15 @@ def test_full_api_mangadex(wait_20s):
 
 
 def test_full_with_input_cbz(wait_20s):
-    url_uuid = "https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie"
+    url_uuid = "https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko"
     language = "en"
     chapters = "1"
     file_format = "cbz"
     download_path = "tests"
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1.cbz")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.cbz"
+    )
     command_args = f"-u {url_uuid} -l {language} -c {chapters} --path {download_path} --format {file_format} --debug --wait 2"
     script_path = "manga-dlp.py"
     os.system(f"python3 {script_path} {command_args}")
@@ -61,13 +65,15 @@ def test_full_with_input_cbz(wait_20s):
 
 
 def test_full_with_input_cbz_info(wait_20s):
-    url_uuid = "https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie"
+    url_uuid = "https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko"
     language = "en"
     chapters = "1"
     file_format = "cbz"
     download_path = "tests"
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1.cbz")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.cbz"
+    )
     command_args = f"-u {url_uuid} -l {language} -c {chapters} --path {download_path} --format {file_format} --wait 2"
     script_path = "manga-dlp.py"
     os.system(f"python3 {script_path} {command_args}")
@@ -82,13 +88,15 @@ def test_full_with_input_cbz_info(wait_20s):
     platform.machine() != "x86_64", reason="pdf only supported on amd64"
 )
 def test_full_with_input_pdf(wait_20s):
-    url_uuid = "https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie"
+    url_uuid = "https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko"
     language = "en"
     chapters = "1"
     file_format = "pdf"
     download_path = "tests"
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1.pdf")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.pdf"
+    )
     command_args = f"-u {url_uuid} -l {language} -c {chapters} --path {download_path} --format {file_format} --debug --wait 2"
     script_path = "manga-dlp.py"
     os.system(f"python3 {script_path} {command_args}")
@@ -100,31 +108,39 @@ def test_full_with_input_pdf(wait_20s):
 
 
 def test_full_with_input_folder(wait_20s):
-    url_uuid = "https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie"
+    url_uuid = "https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko"
     language = "en"
     chapters = "1"
     file_format = ""
     download_path = "tests"
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire"
+    )
+    metadata_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire/ComicInfo.xml"
+    )
     command_args = f"-u {url_uuid} -l {language} -c {chapters} --path {download_path} --format '{file_format}' --debug --wait 2"
     script_path = "manga-dlp.py"
     os.system(f"python3 {script_path} {command_args}")
 
     assert manga_path.exists() and manga_path.is_dir()
     assert chapter_path.exists() and chapter_path.is_dir()
+    assert metadata_path.exists() and metadata_path.is_file()
     # cleanup
     shutil.rmtree(manga_path, ignore_errors=True)
 
 
 def test_full_with_input_skip_cbz(wait_10s):
-    url_uuid = "https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie"
+    url_uuid = "https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko"
     language = "en"
     chapters = "1"
     file_format = "cbz"
     download_path = "tests"
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1.cbz")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.cbz"
+    )
     command_args = f"-u {url_uuid} -l {language} -c {chapters} --path {download_path} --format {file_format} --debug --wait 2"
     script_path = "manga-dlp.py"
     manga_path.mkdir(parents=True, exist_ok=True)
@@ -138,13 +154,15 @@ def test_full_with_input_skip_cbz(wait_10s):
 
 
 def test_full_with_input_skip_folder(wait_10s):
-    url_uuid = "https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie"
+    url_uuid = "https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko"
     language = "en"
     chapters = "1"
     file_format = ""
     download_path = "tests"
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire"
+    )
     command_args = f"-u {url_uuid} -l {language} -c {chapters} --path {download_path} --format '{file_format}' --debug --wait 2"
     script_path = "manga-dlp.py"
     chapter_path.mkdir(parents=True, exist_ok=True)
@@ -156,8 +174,12 @@ def test_full_with_input_skip_folder(wait_10s):
 
     assert chapter_path.is_dir()
     assert found_files == []
-    assert not Path("tests/Shikimori's Not Just a Cutie/Ch. 1.cbz").exists()
-    assert not Path("tests/Shikimori's Not Just a Cutie/Ch. 1.zip").exists()
+    assert not Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.cbz"
+    ).exists()
+    assert not Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.zip"
+    ).exists()
     # cleanup
     shutil.rmtree(manga_path, ignore_errors=True)
 
@@ -168,12 +190,14 @@ def test_full_with_read_cbz(wait_20s):
     chapters = "1"
     file_format = "cbz"
     download_path = "tests"
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1.cbz")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.cbz"
+    )
     command_args = f"--read {str(url_list)} -l {language} -c {chapters} --path {download_path} --format {file_format} --debug --wait 2"
     script_path = "manga-dlp.py"
     url_list.write_text(
-        "https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie"
+        "https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko"
     )
 
     os.system(f"python3 {script_path} {command_args}")
@@ -190,14 +214,16 @@ def test_full_with_read_skip_cbz(wait_10s):
     chapters = "1"
     file_format = "cbz"
     download_path = "tests"
-    manga_path = Path("tests/Shikimori's Not Just a Cutie")
-    chapter_path = Path("tests/Shikimori's Not Just a Cutie/Ch. 1.cbz")
+    manga_path = Path("tests/Tomo-chan wa Onna no ko")
+    chapter_path = Path(
+        "tests/Tomo-chan wa Onna no ko/Ch. 1 - Once In A Life Time Misfire.cbz"
+    )
     command_args = f"--read {str(url_list)} -l {language} -c {chapters} --path {download_path} --format {file_format} --debug --wait 2"
     script_path = "manga-dlp.py"
     manga_path.mkdir(parents=True, exist_ok=True)
     chapter_path.touch()
     url_list.write_text(
-        "https://mangadex.org/title/0aea9f43-d4a9-4bf7-bebc-550a512f9b95/shikimori-s-not-just-a-cutie"
+        "https://mangadex.org/title/76ee7069-23b4-493c-bc44-34ccbf3051a8/tomo-chan-wa-onna-no-ko"
     )
 
     os.system(f"python3 {script_path} {command_args}")